Functions

These functions constitute the main interface for working with OmniSci.

Authentication

OmniSci.connect — Function.

connect(host::String, port::Int, user::String, passwd::String, dbname::String)

Connect to an OmniSci database.

Examples

julia> conn = connect("localhost", 9091, "mapd", "HyperInteractive", "mapd")
Connected to localhost:9091

source

OmniSci.disconnect — Function.

disconnect(conn::OmniSciConnection)

Close connection to OmniSci database.

Examples

julia> disconnect(conn)
Connection to localhost:9091 closed

source

Querying Data

OmniSci.sql_execute_df — Function.

sql_execute_df(conn::OmniSciConnection, query::String, device_type::Int, device_id::Int, first_n::Int = -1)

source

OmniSci.sql_execute_gdf — Function.

sql_execute_gdf(conn::OmniSciConnection, query::String, device_id::Int, first_n::Int = -1)

source

OmniSci.sql_execute — Function.

sql_execute(conn::OmniSciConnection, query::String, first_n::Int = -1, at_most_n::Int = -1)

source

Loading Data

OmniSci.create_table — Function.

create_table(conn::OmniSciConnection, table_name::String, row_desc::TRowDescriptor, table_type::TTableType.Enum)

source

OmniSci.load_table — Function.

load_table(conn::OmniSciConnection, table_name::String, rows::Vector{TStringRow})

source

OmniSci.load_table_binary_arrow — Function.

load_table_binary_arrow(conn::OmniSciConnection, table_name::String, arrow_stream::Vector{UInt8})

source

OmniSci.load_table_binary — Function.

load_table_binary(conn::OmniSciConnection, table_name::String, rows::Vector{TRow})

source

OmniSci.load_table_binary_columnar — Function.

load_table_binary_columnar(conn::OmniSciConnection, table_name::String, cols::Vector{TColumn})

source

OmniSci.import_geo_table — Function.

import_geo_table(conn::OmniSciConnection, table_name::String, file_name::String, copy_params::TCopyParams, row_desc::TRowDescriptor)

source

Table Metadata

OmniSci.get_tables_meta — Function.

get_tables_meta(conn::OmniSciConnection, as_df::Bool = true)

Get metadata for tables in database specified in connect().

Examples

julia> metad = get_tables_meta(conn)
5×6 DataFrame. Omitted printing of 1 columns
│ Row │ is_replicated │ is_view │ max_rows            │ num_cols │ shard_count │
│     │ Bool          │ Bool    │ Int64               │ Int64    │ Int64       │
├─────┼───────────────┼─────────┼─────────────────────┼──────────┼─────────────┤
│ 1   │ false         │ false   │ 4611686018427387904 │ 4        │ 0           │
│ 2   │ false         │ false   │ 4611686018427387904 │ 6        │ 0           │
│ 3   │ false         │ false   │ 4611686018427387904 │ 64       │ 0           │
│ 4   │ false         │ false   │ 4611686018427387904 │ 56       │ 0           │
│ 5   │ false         │ false   │ 4611686018427387904 │ 42       │ 0           │

source

OmniSci.get_table_details — Function.

get_table_details(conn::OmniSciConnection, table_name::String)

Get table details such as column names and types.

Examples

julia> tbl_detail = get_table_details(conn, "mapd_states")
TTableDetails(TColumnType[TColumnType("id", TTypeInfo(6, 4, true, false, 0, 0, 32, -1), false, "", false, false), TColumnType("abbr", TTypeInfo(6, 4, true, false, 0, 0, 32, -1), false, "", false, false), TColumnType("name", TTypeInfo(6, 4, true, false, 0, 0, 32, -1), false, "", false, false), TColumnType("mapd_geo", TTypeInfo(16, 6, true, false, 23, 4326, 32, -1), false, "", false, false)], 32000000, 2097152, 4611686018427387904, "", 0, "[]", false, 0)

source

Dashboards

OmniSci.create_dashboard — Function.

create_dashboard(conn::OmniSciConnection, dashboard_name::String, dashboard_state::String, image_hash::String, dashboard_metadata::String)

source

OmniSci.share_dashboard — Function.

share_dashboard(conn::OmniSciConnection, dashboard_id::Integer, groups::Vector{String}, objects::Vector{String}, permissions::TDashboardPermissions)

source

OmniSci.unshare_dashboard — Function.

unshare_dashboard(conn::OmniSciConnection, dashboard_id::Integer, groups::Vector{String}, objects::Vector{String}, permissions::TDashboardPermissions)

source

OmniSci.get_dashboards — Function.

get_dashboards(conn::OmniSciConnection, as_df::Bool = true)

Gets dashboards that user submitted during connect() can access.

Examples

julia> getdbs = get_dashboards(conn)
7×8 DataFrame. Omitted printing of 3 columns
│ Row │ dashboard_id │ dashboard_metadata │ dashboard_name │ dashboard_owner │ dashboard_state │
│     │ Int32        │ String             │ String         │ String          │ String          │
├─────┼──────────────┼────────────────────┼────────────────┼─────────────────┼─────────────────┤
│ 1   │ 9            │ metadata           │ 0vcAQEO1ZD     │ mapd            │                 │
│ 2   │ 6            │ metadata           │ QI0JsthBsB     │ mapd            │                 │
│ 3   │ 5            │ metadata           │ Srm72rCJHa     │ mapd            │                 │
│ 4   │ 4            │ metadata           │ sO0XgMUOZH     │ mapd            │                 │
│ 5   │ 1            │ metadata           │ testdash       │ mapd            │                 │
│ 6   │ 2            │ metadata           │ testdash2      │ mapd            │                 │
│ 7   │ 3            │ metadata           │ testdash3      │ mapd            │                 │

source

OmniSci.get_dashboard_grantees — Function.

get_dashboard_grantees(conn::OmniSciConnection, dashboard_id::Integer)

source

OmniSci.replace_dashboard — Function.

replace_dashboard(conn::OmniSciConnection, dashboard_id::Integer, dashboard_name::String, dashboard_owner::String, dashboard_state::String, image_hash::String, dashboard_metadata::String)

source

OmniSci.delete_dashboard — Function.

delete_dashboard(conn::OmniSciConnection, dashboard_id::Integer)

source

OmniSci.render_vega — Function.

render_vega(conn::OmniSciConnection, widget_id::Int, vega_json::String, compression_level::Int)

source

Metadata

OmniSci.get_databases — Function.

get_databases(conn::OmniSciConnection, as_df::Bool=true)

Get list of databases.

Examples

julia> db = get_databases(conn)
1×2 DataFrame
│ Row │ db_name │ db_owner │
│     │ String  │ String   │
├─────┼─────────┼──────────┤
│ 1   │ mapd    │ mapd     │

source

OmniSci.get_users — Function.

get_users(conn::OmniSciConnection, as_df::Bool = true)

Get list of users who have access to database specified in connect().

Examples

julia> users = get_users(conn)
1×1 DataFrame
│ Row │ users  │
│     │ String │
├─────┼────────┤
│ 1   │ mapd   │

source

OmniSci.get_roles — Function.

get_roles(conn::OmniSciConnection, as_df::Bool = true)

source

OmniSci.get_hardware_info — Function.

get_hardware_info(conn::OmniSciConnection)

Displays selected properties of hardware where OmniSci running, such as GPU and CPU information.

Examples

julia> hardware = get_hardware_info(conn)
TClusterHardwareInfo(THardwareInfo[THardwareInfo(0, 12, 0, 0, "", TGpuSpecification[])])

source

OmniSci.get_status — Function.

get_status(conn::OmniSciConnection)

Displays properties of OmniSci server, such as version and rendering capabilities.

Examples

julia> status = get_status(conn)
1-element Array{TServerStatus,1}:
 TServerStatus(false, "4.2.0dev-20181003-0206b9f92c", false, 1539095178, "ee", "aggregator", false)

source

OmniSci.get_all_roles_for_user — Function.

get_all_roles_for_user(conn::OmniSciConnection, userName::String)

source

OmniSci.get_db_objects_for_grantee — Function.

get_db_objects_for_grantee(conn::OmniSciConnection, roleName::String)

source

OmniSci.get_db_object_privs — Function.

get_db_object_privs(conn::OmniSciConnection, objectName::String, type_::Integer)

source