Internal

Internal Functions

For completeness of wrapping the OmniSci Thrift interface, the following functions have been implemented. These functions represent either internal OmniSci functionality and/or functions that have other convenience methods defined around them.

It is not expected that a user would need to use these functions under normal circumstances.

Licensing

OmniSci.get_license_claims — Function.

get_license_claims(conn::OmniSciConnection)

source

OmniSci.set_license_key — Function.

set_license_key(conn::OmniSciConnection, key::String)

source

Memory Management

deallocate_df
clear_gpu_memory
clear_cpu_memory
set_execution_mode
get_memory

Misc

sql_validate
import_table_status
interrupt
detect_column_types
render_vega