Functions

These functions constitute the main interface for working with OmniSci.

Authentication

OmniSci.connect — Function.

connect(host::String, port::Int, user::String, passwd::String, dbname::String)

Connect to an OmniSci database.

Examples

julia> conn = connect("localhost", 9091, "mapd", "HyperInteractive", "mapd")
Connected to localhost:9091

source

OmniSci.disconnect — Function.

disconnect(conn::OmniSciConnection)

Close connection to OmniSci database.

Examples

julia> disconnect(conn)
Connection to localhost:9091 closed

source

Querying Data

OmniSci.sql_execute_df — Function.

sql_execute_df(conn::OmniSciConnection, query::String, device_type::Int, device_id::Int, first_n::Int = -1)

source

OmniSci.sql_execute_gdf — Function.

sql_execute_gdf(conn::OmniSciConnection, query::String, device_id::Int, first_n::Int = -1)

source

OmniSci.sql_execute — Function.

sql_execute(conn::OmniSciConnection, query::String, column_format::Bool = true, first_n::Int = -1, at_most_n::Int = -1)

source

Loading Data

OmniSci.create_table — Function.

create_table(conn::OmniSciConnection, table_name::String, row_desc::TRowDescriptor, table_type::TTableType.Enum)

source

OmniSci.load_table — Function.

load_table(conn::OmniSciConnection, table_name::String, rows::Vector{TStringRow})

source

OmniSci.load_table_binary_arrow — Function.

load_table_binary_arrow(conn::OmniSciConnection, table_name::String, arrow_stream::Vector{UInt8})

source

OmniSci.load_table_binary — Function.

load_table_binary(conn::OmniSciConnection, table_name::String, rows::Vector{TRow})

source

OmniSci.load_table_binary_columnar — Function.

load_table_binary_columnar(conn::OmniSciConnection, table_name::String, cols::Vector{TColumn})

source

OmniSci.import_geo_table — Function.

import_geo_table(conn::OmniSciConnection, table_name::String, file_name::String, copy_params::TCopyParams, row_desc::TRowDescriptor)

source

OmniSci.import_table — Function.

import_table(conn::OmniSciConnection, table_name::String, file_name::String, copy_params::TCopyParams)

source

Table Metadata

OmniSci.get_tables — Function.

get_tables(conn::OmniSciConnection)

Get tables and views for authenticated database specified in connect().

Examples

julia> tbl = get_tables(conn)
4-element Array{String,1}:
 "mapd_states"
 "mapd_counties"
 "mapd_countries"
 "nyc_trees_2015_683k"

source

OmniSci.get_physical_tables — Function.

get_physical_tables(conn::OmniSciConnection)

Get tables for authenticated database specified in connect().

Examples

julia> ptbl = get_physical_tables(conn)
4-element Array{String,1}:
 "mapd_states"
 "mapd_counties"
 "mapd_countries"
 "nyc_trees_2015_683k"

source

OmniSci.get_tables_meta — Function.

get_tables_meta(conn::OmniSciConnection)

Get metadata for tables in database specified in connect().

Examples

julia> metad = get_tables_meta(conn)
4-element Array{TTableMeta,1}:
 TTableMeta("mapd_states", 4, Int32[6, 16], false, false, 0, 4611686018427387904)
 TTableMeta("mapd_counties", 6, Int32[6, 16], false, false, 0, 4611686018427387904)
 TTableMeta("mapd_countries", 64, Int32[1, 5, 6, 16], false, false, 0, 4611686018427387904)
 TTableMeta("nyc_trees_2015_683k", 42, Int32[0, 3, 6, 9], false, false, 0, 4611686018427387904)

source

OmniSci.get_views — Function.

get_views(conn::OmniSciConnection)

Get views for authenticated database specified in connect().

Examples

julia> vw = get_views(conn)
0-element Array{String,1}

source

OmniSci.get_table_details — Function.

get_table_details(conn::OmniSciConnection, table_name::String)

Get table details such as column names and types.

Examples

julia> tbl_detail = get_table_details(conn, "mapd_states")
TTableDetails(TColumnType[TColumnType("id", TTypeInfo(6, 4, true, false, 0, 0, 32, -1), false, "", false, false), TColumnType("abbr", TTypeInfo(6, 4, true, false, 0, 0, 32, -1), false, "", false, false), TColumnType("name", TTypeInfo(6, 4, true, false, 0, 0, 32, -1), false, "", false, false), TColumnType("mapd_geo", TTypeInfo(16, 6, true, false, 23, 4326, 32, -1), false, "", false, false)], 32000000, 2097152, 4611686018427387904, "", 0, "[]", false, 0)

source

Dashboards

OmniSci.create_dashboard — Function.

create_dashboard(conn::OmniSciConnection, dashboard_name::String, dashboard_state::String, image_hash::String, dashboard_metadata::String)

source

OmniSci.share_dashboard — Function.

share_dashboard(conn::OmniSciConnection, dashboard_id::Integer, groups::Vector{String}, objects::Vector{String}, permissions::TDashboardPermissions)

source

OmniSci.get_dashboard — Function.

get_dashboard(conn::OmniSciConnection, dashboard_id::Integer)

source

OmniSci.unshare_dashboard — Function.

unshare_dashboard(conn::OmniSciConnection, dashboard_id::Integer, groups::Vector{String}, objects::Vector{String}, permissions::TDashboardPermissions)

source

OmniSci.get_dashboards — Function.

get_dashboards(conn::OmniSciConnection)

source

OmniSci.get_dashboard_grantees — Function.

get_dashboard_grantees(conn::OmniSciConnection, dashboard_id::Integer)

source

OmniSci.replace_dashboard — Function.

replace_dashboard(conn::OmniSciConnection, dashboard_id::Integer, dashboard_name::String, dashboard_owner::String, dashboard_state::String, image_hash::String, dashboard_metadata::String)

source

OmniSci.delete_dashboard — Function.

delete_dashboard(conn::OmniSciConnection, dashboard_id::Integer)

source

Metadata

OmniSci.get_databases — Function.

get_databases(conn::OmniSciConnection)

Get list of databases.

Examples

julia> db = get_databases(conn)
1-element Array{TDBInfo,1}:
 TDBInfo("mapd", "mapd")

source

OmniSci.get_users — Function.

get_users(conn::OmniSciConnection)

Get list of users who have access to database specified in connect().

Examples

julia> users = get_users(conn)
1-element Array{String,1}:
 "mapd"

source

OmniSci.get_roles — Function.

get_roles(conn::OmniSciConnection)

source

OmniSci.get_hardware_info — Function.

get_hardware_info(conn::OmniSciConnection)

Displays selected properties of hardware where OmniSci running, such as GPU and CPU information.

Examples

julia> hardware = get_hardware_info(conn)
TClusterHardwareInfo(THardwareInfo[THardwareInfo(0, 12, 0, 0, "", TGpuSpecification[])])

source

OmniSci.get_status — Function.

get_status(conn::OmniSciConnection)

Displays properties of OmniSci server, such as version and rendering capabilities.

Examples

julia> status = get_status(conn)
1-element Array{TServerStatus,1}:
 TServerStatus(false, "4.2.0dev-20181003-0206b9f92c", false, 1539095178, "ee", "aggregator", false)

source

OmniSci.get_all_roles_for_user — Function.

get_all_roles_for_user(conn::OmniSciConnection, userName::String)

source

OmniSci.get_db_objects_for_grantee — Function.

get_db_objects_for_grantee(conn::OmniSciConnection, roleName::String)

source

OmniSci.get_version — Function.

get_version(conn::OmniSciConnection)

Get OmniSci software version.

Examples

julia> version = get_version(conn)
"4.2.0dev-20181003-0206b9f92c"

source

OmniSci.get_db_object_privs — Function.

get_db_object_privs(conn::OmniSciConnection, objectName::String, type_::Integer)

source