Functions

These functions constitute the main interface for working with OmniSci. If a function is not exported for the package, the assumption is that an end-user shouldn't need to use it. If you find otherwise, please open an issue for discussion.

Authentication

OmniSci.connect — Function.

connect(host::String, port::Int, user::String, passwd::String, dbname::String)

Connect to an OmniSci database.

Examples

julia> conn = connect("localhost", 9091, "mapd", "HyperInteractive", "mapd")
Connected to localhost:9091

source

OmniSci.disconnect — Function.

disconnect(conn::OmniSciConnection)

Close connection to OmniSci database.

Examples

julia> disconnect(conn)
Connection to localhost:9091 closed

source

Querying Data

OmniSci.sql_execute — Function.

sql_execute(conn::OmniSciConnection, query::String; first_n::Int = -1, at_most_n::Int = -1, as_df::Bool = true)

Execute a SQL query.

source

OmniSci.sql_execute_df — Function.

sql_execute_df(conn::OmniSciConnection, query::String, device_type::Int, device_id::Int, first_n::Int = -1)

Execute a SQL query using Apache Arrow IPC (CPU). This method requires running the code in the same environment where OmniSci is running.

source

OmniSci.sql_execute_gdf — Function.

sql_execute_gdf(conn::OmniSciConnection, query::String, device_id::Int; first_n::Int = -1)

Execute a SQL query using Apache Arrow IPC (GPU). This method requires running the code in the same environment where OmniSci is running.

source

Loading Data

OmniSci.create_table — Function.

create_table(conn::OmniSciConnection, table_name::String, df::DataFrame; dryrun::Bool = false, precision::Tuple{Int,Int} = (0,0))

Create a table in the database specified during authentication. This method takes a Julia DataFrame, reads the column types and creates the equivalent create table statement, optionally printing the create table statement instead of executing the statement using sql_execute method.

Note: This method is for convenience purposes only! It does not guarantee an optimized table statement. Additionally, if Decimal columns present in 'df', user must set precision/scale via the 'precision' argument.

Examples

julia> create_table(conn, "test", df)

source

OmniSci.load_table — Function.

load_table(conn::OmniSciConnection, table_name::String, rows::Vector{TStringRow})

source

load_table(conn::OmniSciConnection, table_name::String, rows::DataFrame)

Load a dataframe into OmniSci. This method loads data row-wise, converting data elements to string before upload. Currently, this method requires the table to already exist on OmniSci.

Examples

julia> load_table(conn, "test", df)

source

OmniSci.load_table_binary_columnar — Function.

load_table_binary_columnar(conn::OmniSciConnection, table_name::String, cols::Vector{TColumn})

source

load_table_binary_columnar(conn::OmniSciConnection, table_name::String, df::DataFrame)

source

OmniSci.load_table_binary_arrow — Function.

load_table_binary_arrow(conn::OmniSciConnection, table_name::String, arrow_stream::Vector{UInt8})

source

Dashboards

OmniSci.get_dashboards — Function.

get_dashboards(conn::OmniSciConnection; as_df::Bool = true)

Gets dashboards that user submitted during connect() can access.

Examples

julia> getdbs = get_dashboards(conn)
7×8 DataFrame. Omitted printing of 3 columns
│ Row │ dashboard_id │ dashboard_metadata │ dashboard_name │ dashboard_owner │ dashboard_state │
│     │ Int32        │ String             │ String         │ String          │ String          │
├─────┼──────────────┼────────────────────┼────────────────┼─────────────────┼─────────────────┤
│ 1   │ 9            │ metadata           │ 0vcAQEO1ZD     │ mapd            │                 │
│ 2   │ 6            │ metadata           │ QI0JsthBsB     │ mapd            │                 │
│ 3   │ 5            │ metadata           │ Srm72rCJHa     │ mapd            │                 │
│ 4   │ 4            │ metadata           │ sO0XgMUOZH     │ mapd            │                 │
│ 5   │ 1            │ metadata           │ testdash       │ mapd            │                 │
│ 6   │ 2            │ metadata           │ testdash2      │ mapd            │                 │
│ 7   │ 3            │ metadata           │ testdash3      │ mapd            │                 │

source

OmniSci.get_dashboard_grantees — Function.

get_dashboard_grantees(conn::OmniSciConnection, dashboard_id::Integer)

source

OmniSci.render_vega — Function.

render_vega(conn::OmniSciConnection, vega_json::String, compression_level::Int = 0)

Render an OmniSci-flavored Vega specification using the backend rendering engine. Note that OmniSci does not currently support the full Vega specification; this method is mostly useful for rendering choropleths and related geospatial charts.

compression_level ranges from 0 (low compression, faster) to 9 (high compression, slower).

Examples

julia> vg = {"width" : 1024, "height" : 1024...}

julia> vega_json = render_vega(conn, vg)

source

Metadata

OmniSci.get_tables_meta — Function.

get_tables_meta(conn::OmniSciConnection; as_df::Bool = true)

Get metadata for tables in database specified in connect().

Examples

julia> metad = get_tables_meta(conn)
5×6 DataFrame
│ Row │ is_replicated │ is_view │ max_rows            │ num_cols │ shard_count │ table_name        │
│     │ Bool          │ Bool    │ Int64               │ Int64    │ Int64       │ String            │
├─────┼───────────────┼─────────┼─────────────────────┼──────────┼─────────────┼───────────────────┤
│ 1   │ false         │ false   │ 4611686018427387904 │ 4        │ 0           │ omnisci_states    │
│ 2   │ false         │ false   │ 4611686018427387904 │ 6        │ 0           │ omnisci_counties  │
│ 3   │ false         │ false   │ 4611686018427387904 │ 64       │ 0           │ omnisci_countries │
│ 4   │ false         │ false   │ 4611686018427387904 │ 4        │ 0           │ test2             │
│ 5   │ false         │ false   │ 4611686018427387904 │ 4        │ 0           │ test              │

source

OmniSci.get_table_details — Function.

get_table_details(conn::OmniSciConnection, table_name::String; as_df::Bool = true)

Get table details such as column names and types.

Examples

julia> tbl_detail = get_table_details(conn, "omnisci_states")
4×21 DataFrame. Omitted printing of 15 columns
│ Row │ col_name    │ col_type     │ comp_param │ encoding │ is_array │ is_physical │
│     │ String      │ DataType     │ Int32      │ String   │ Bool     │ Bool        │
├─────┼─────────────┼──────────────┼────────────┼──────────┼──────────┼─────────────┤
│ 1   │ id          │ String       │ 32         │ Dict     │ false    │ false       │
│ 2   │ abbr        │ String       │ 32         │ Dict     │ false    │ false       │
│ 3   │ name        │ String       │ 32         │ Dict     │ false    │ false       │
│ 4   │ omnisci_geo │ MultiPolygon │ 32         │ GeoInt   │ false    │ false       │

source

OmniSci.get_databases — Function.

get_databases(conn::OmniSciConnection; as_df::Bool=true)

Get list of databases.

Examples

julia> db = get_databases(conn)
1×2 DataFrame
│ Row │ db_name │ db_owner │
│     │ String  │ String   │
├─────┼─────────┼──────────┤
│ 1   │ mapd    │ mapd     │

source

OmniSci.get_users — Function.

get_users(conn::OmniSciConnection; as_df::Bool = true)

Get list of users who have access to database specified in connect().

Examples

julia> users = get_users(conn)
1×1 DataFrame
│ Row │ users  │
│     │ String │
├─────┼────────┤
│ 1   │ mapd   │

source

OmniSci.get_roles — Function.

get_roles(conn::OmniSciConnection; as_df::Bool = true)

source

OmniSci.get_hardware_info — Function.

get_hardware_info(conn::OmniSciConnection)

Displays selected properties of hardware where OmniSci running, such as GPU and CPU information.

Examples

julia> hardware = get_hardware_info(conn)
OmniSci.TClusterHardwareInfo(OmniSci.THardwareInfo[THardwareInfo(0, 12, 0, 0, "", OmniSci.TGpuSpecification[])])

source

OmniSci.get_status — Function.

get_status(conn::OmniSciConnection)

Displays properties of OmniSci server, such as version and rendering capabilities.

Examples

julia> status = get_status(conn)
OmniSci.TServerStatus

  read_only: false
  version: 4.1.3-20180926-66c2aee949
  rendering_enabled: false
  start_time: 1540579280
  edition: ce
  host_name: aggregator
  poly_rendering_enabled: false

source

OmniSci.get_all_roles_for_user — Function.

get_all_roles_for_user(conn::OmniSciConnection, userName::String; as_df::Bool = true)

source