Internal

Internal Functions

For completeness of wrapping the OmniSci Thrift interface, the following functions have been implemented. These functions represent either internal OmniSci functionality and/or functions that have other convenience methods defined around them.

It is not expected that a user would need to use these functions under normal circumstances.

Licensing

OmniSci.get_license_claims — Function.

get_license_claims(conn::OmniSciConnection)

source

OmniSci.set_license_key — Function.

set_license_key(conn::OmniSciConnection, key::String)

source

Memory Management

OmniSci.deallocate_df — Function.

deallocate_df(conn::OmniSciConnection, df::TDataFrame, device_type::Int, device_id::Int)

source

OmniSci.clear_gpu_memory — Function.

clear_gpu_memory(conn::OmniSciConnection)

source

OmniSci.clear_cpu_memory — Function.

clear_cpu_memory(conn::OmniSciConnection)

source

OmniSci.set_execution_mode — Function.

set_execution_mode(conn::OmniSciConnection, mode::TExecuteMode.Enum)

source

OmniSci.get_memory — Function.

get_memory(conn::OmniSciConnection, memory_level::String)

source

Misc

OmniSci.sql_validate — Function.

sql_validate(conn::OmniSciConnection, query::String)

source

OmniSci.import_table_status — Function.

import_table_status(conn::OmniSciConnection, import_id::String)

source

OmniSci.interrupt — Function.

interrupt(conn::OmniSciConnection)

source

OmniSci.detect_column_types — Function.

detect_column_types(conn::OmniSciConnection, file_name::String, copy_params::TCopyParams)

source

OmniSci.render_vega — Function.

render_vega(conn::OmniSciConnection, widget_id::Int, vega_json::String, compression_level::Int)

source